requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dotenv-python==0.0.1
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.18.0
MarkupSafe==1.1.1
requests==2.25.1
urllib3==1.26.2
Werkzeug==1.0.1
